+++ tests/Cooked/WithdrawalsSpec.hs
@@ -1,32 +1,9 @@
 module Cooked.WithdrawalsSpec where
 
 import Cooked
-import Plutus.Script.Utils.V3 qualified as Script
+import Plutus.Withdrawals
 import PlutusLedgerApi.V3 qualified as Api
-import PlutusTx qualified
-import PlutusTx.AssocMap qualified as PMap
-import PlutusTx.Prelude qualified as PlutusTx
 import Test.Tasty
-
-{-# INLINEABLE checkWithdrawalPurpose #-}
-checkWithdrawalPurpose :: Script.RewardingPurposeType' Integer Api.TxInfo
-checkWithdrawalPurpose cred quantity (Api.TxInfo {txInfoWdrl}) =
-  case PMap.toList txInfoWdrl of
-    [(cred', Api.Lovelace n)] ->
-      if cred PlutusTx.== cred'
-        then (n PlutusTx.== quantity) || PlutusTx.traceError "Wrong quantity."
-        else PlutusTx.traceError "Wrong credential."
-    _ -> PlutusTx.traceError "Wrong withdrawal."
-
-checkWithdrawalVersionedScript :: Script.Versioned Script.Script
-checkWithdrawalVersionedScript =
-  Script.toVersioned $
-    Script.MultiPurposeScript @() $
-      Script.toScript $$(PlutusTx.compile [||script||])
-  where
-    script =
-      Script.mkMultiPurposeScript $
-        Script.falseTypedMultiPurposeScript `Script.withRewardingPurpose` checkWithdrawalPurpose
 
 testWithdrawingScript :: (MonadBlockChain m) => Integer -> Integer -> m ()
 testWithdrawingScript n1 n2 =
