+++ tests/Plutus/ProposingScript.hs
@@ -4,49 +4,45 @@ module Plutus.ProposingScript where
 
 import Plutus.Script.Utils.V3 qualified as Script
 import PlutusLedgerApi.V3 qualified as Api
-import PlutusTx.AssocMap qualified as PMap
-import PlutusTx.Builtins qualified as PlutusTx hiding (head)
-import PlutusTx.Eq qualified as PlutusTx
-import PlutusTx.IsData qualified as PlutusTx
-import PlutusTx.Prelude qualified as PlutusTx
-import PlutusTx.TH qualified as PlutusTx
-import Prelude qualified as HS
+import PlutusTx
+import PlutusTx.AssocMap qualified as Map
+import PlutusTx.Prelude
 
 {-# INLINEABLE checkParameterChangeProposingPurpose #-}
 checkParameterChangeProposingPurpose :: Script.ProposingPurposeType ()
 checkParameterChangeProposingPurpose _ (Api.ProposalProcedure _ _ (Api.ParameterChange _ (Api.ChangedParameters dat) _)) _ _ =
-  let innerMap = PlutusTx.unsafeFromBuiltinData @(PMap.Map PlutusTx.Integer PlutusTx.Integer) dat
-   in ((PMap.toList innerMap PlutusTx.== [(0, 100)]) PlutusTx.|| PlutusTx.traceError "wrong map")
-checkParameterChangeProposingPurpose _ _ _ _ = PlutusTx.traceError "Wrong proposal procedure"
+  let innerMap = unsafeFromBuiltinData @(Map.Map Integer Integer) dat
+   in ((Map.toList innerMap == [(0, 100)]) || traceError "wrong map")
+checkParameterChangeProposingPurpose _ _ _ _ = traceError "Wrong proposal procedure"
 
 checkProposingScript :: Script.Versioned Script.Script
 checkProposingScript =
-  Script.toVersioned HS.$
-    Script.MultiPurposeScript @() HS.$
-      Script.toScript $$(PlutusTx.compile [||script||])
+  Script.toVersioned
+    $ Script.MultiPurposeScript @()
+    $ Script.toScript $$(compile [||script||])
   where
     script =
-      Script.mkMultiPurposeScript HS.$
-        Script.falseTypedMultiPurposeScript
-          `Script.withProposingPurpose` checkParameterChangeProposingPurpose
+      Script.mkMultiPurposeScript
+        $ Script.falseTypedMultiPurposeScript
+        `Script.withProposingPurpose` checkParameterChangeProposingPurpose
 
 -- | A dummy false proposing validator
 alwaysFalseProposingValidator :: Script.Versioned Script.Script
 alwaysFalseProposingValidator =
-  Script.toVersioned HS.$
-    Script.MultiPurposeScript @() HS.$
-      Script.toScript $$(PlutusTx.compile [||script||])
+  Script.toVersioned
+    $ Script.MultiPurposeScript @()
+    $ Script.toScript $$(compile [||script||])
   where
     script = Script.mkMultiPurposeScript Script.falseTypedMultiPurposeScript
 
 -- | A dummy true proposing validator
 alwaysTrueProposingValidator :: Script.Versioned Script.Script
 alwaysTrueProposingValidator =
-  Script.toVersioned HS.$
-    Script.MultiPurposeScript @() HS.$
-      Script.toScript $$(PlutusTx.compile [||script||])
+  Script.toVersioned
+    $ Script.MultiPurposeScript @()
+    $ Script.toScript $$(compile [||script||])
   where
     script =
-      Script.mkMultiPurposeScript HS.$
-        Script.falseTypedMultiPurposeScript
-          `Script.withProposingPurpose` (\_ _ () () -> PlutusTx.True)
+      Script.mkMultiPurposeScript
+        $ Script.falseTypedMultiPurposeScript
+        `Script.withProposingPurpose` (\_ _ () () -> True)
